src/common/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.1
httpx==0.25.1
python-dotenv==1.0.0
pyjwt==2.8.0
retry==0.9.2
python-json-logger==2.0.7
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
redis==5.0.1
aioredis==2.0.1 